+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.12.0 (2019-05-07)
 
-* Change Script Button to download a Python script.
+* Change Script button to download a Python script.
+* Change Script button to not require logging in (since the script does that).
 
 # v0.11.0 (2019-04-30)